fix: rebase and rename package to sign
* Rebased branch to master and resolved merge conflicts.
* Renamed package gpg_signing to sign to allow for additional signing methods.
* Refactored code for readability.
